Ensure channel names and client_ids are always UTF-8
Ruby 1.9 uses ASCII_8BIT encoding by default.  This causes a lot of problems because the Ruby 2+ convention is that ASCII_8BIT is used to represent binary data.  Therefore, we need to reject ASCII_8BIT for channel names and client_ids to prevent them from being seen as binary data when packed using MessagePack.