Refactor #once_state_changed into StateEmitter
This method belongs to StateEmitter, not the Connection object.  Now that commit b5c70c2 “Allow changes to EventEmitter callbacks from within callbacks without side effects” is implemented, EventMachine is not needed to defer changes to the callback until the next tick to avoid side effects on the #trigger method.