Simplify RSpec “hack” to support EventMachine
Previously when an expectation was failing, the after callbacks were executed but the exception was not passed to the outer :after blocks.  As such, rspec-retry would not retry those EM failing tests.  This fixes that and ensures the code block within the test has the EM reactor wrapping it without having to explicitly wrap all EM tests thus reducing noise.