Optimisation: Log with blocks instead of interpolated strings
Previously if an exception was raised in an interpolated value in a log message such as “error: ${this_will_raise_an_error}” then the execution would stop at that point which is not the intention given it’s just logging.

Separately, passing interpolated strings to the logger when most of the time the logging is not outputted (such as debug logging) is expensive.  This ensures that expensive serialisation operations within interpolated vars is only done when needed.

Fixes https://github.com/ably/ably-ruby/issues/107