Reduce log failure noise (with a hack)
RSpec assertion failures are now caught and the verbose logging that is included in each test now only includes the logs up until the point of failure. Whilst this is hacky, it won’t break anything if RSpec in the future changes its internals, we will simply have more noisy logs again.