Consistent use of deferrable errback for failures for the Realtime lib
If a method is invoked and it’s failure / success is determined by the client library’s state, then the deferrable errback should always be used i.e. a developer can then reliably always use errback and callback to deal with failure & success respectively.

Exceptions are now only raised when a method is called with invalid arguments.