Simplify msgSerial handling
There is no need to rewind msgSerial because a connection recovers or not successfully.  Rolling back is dangerous because of queued messages. Instead, operate on the basis that if a msgSerial is provided in a recovery key, then use that msgSerial regardless of whether that connection recovers or not.  This is the safest approach ensuring messages are always published when the connection is established, and not de-duped because of existing serial numbers.