Turn off continue-on-error in CI
We want the workflow run to fail if any of the jobs fails.

The motivation for the current configuration is not clear to me, but I’m
guessing the idea was that the failure of a single job shouldn’t cancel
the others. Turning off fail-fast should do the same thing.

(Copied from ably-js commit ddef9f7.)