spec: Don't expect integers to round-trip through msg.extras
Internally, we store msg.extras in a protobuf.Struct which stores a JSON
number in a double field:

https://github.com/protocolbuffers/protobuf/blob/5cbf13bc9ab0f883bcb664241f6be8d962915569/src/google/protobuf/struct.proto#L56-L78

This means that when a connection is using binary encoding (i.e.
msgpack), any numbers in msg.extras are encoded as floats even if the
inbound message encoded the number as an int type, so the inbound and
outbound representations are not guaranteed to be the same, although the
semantic numeric value will be the same.

This commit handles this nuance by not expecting an integer to come back
as an integer in msg.extras, but expecting a float to instead.

Signed-off-by: Lewis Marshall <lewis.marshall@ably.com>