Don't use fallback hosts for Realtime token errors
This is a partial fix for #444 (nature of error not considered when
deciding whether to use a fallback host for realtime connections).

Here, I implement just enough of RTN17f to make sure that we don't use a
fallback host in response to a token error. This fixes an intermittent
failure in the test case on connection_failures_spec.rb:1361, which
tests that we correctly handle token errors when resuming a connection
per RTN15c5. This test was intermittently failing because it would
incorrectly try and use a fallback host after receiving the token error,
causing the resume attempt to sometimes fail with "Unable to resume
connection from another site".

I don't have loads of time on to spend on this issue right now, which is
why I'm applying the smallest possible fix. We need to return to #444
and properly consider how to identify the types of errors mentioned in
RTN17f.